refactor(ui): split admin disease browser components
  - extract search form, labels, group option mapping, and result views
    into focused internal modules
  - reduce oversized disease page and results files while preserving
    behavior

  Refs: BEJ-123
